backend/pyproject.toml
[project]
name = "ats-backend"
version = "0.1.0"
description = "Auto Trading System Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.2.0",
    "pyyaml>=6.0.2",
    "mt5linux>=0.0.8",
    "rpyc>=5.2,<6",
    "watchdog>=6.0.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
